Rename QuizzerStore to DeckStudyStore.
Make Build & Study pages independent from Deck Master Index page; this makes `mode` & `step` props unnecessary. Make Quizzer Show page the Study page in the Deck Master, with former Study page (Card Review) becoming the Practice page, parallel to the Quiz page. DeckActions "Study Deck" option redirects to Study page.

Remove Quizzer Index page; it's equivalent to the Deck Master Index (Select) page. Remove Quizzer item in NavSidebar.

Remove QuizzerController & move logic to DeckMasterController; `generate()` method becomes `getQuiz()`; `show()` becomes `study()`; former `study()` method is replaced by `getCards()`. Remove Quizzer routes & merge into DeckMaster, without scorable wildcards.

Quiz generation in QuizService always takes Deck models.

Condition viewing Scores & generating Quizzes on Student role.
