+++ resources/js/Pages/Academy/Scores/History.vue
@@ -112,10 +112,7 @@ watch(() => props.selectedScore, (newVal) => {
                             scorable_type
                         }}</span>
                         yet. When you do, you will see a history of your Scores here.
-                        <Link :href="route('quizzer.show', { scorable_type: scorable_type, scorable_id: model.id })">Go
-                            to
-                            the Quizzer.
-                        </Link>
+                        <Link :href="route('deck-master.study', model.id)">Go to the Quizzer.</Link>
                     </p>
                 </AppTip>
                 <template v-if="totalCount > 0">
