+++ resources/js/Pages/Workbench/DeckMaster/Index.vue
@@ -1,26 +1,25 @@
 <script setup>
-import {onMounted, ref} from "vue";
+import {onBeforeUnmount, onMounted, ref} from "vue";
 import Layout from "../../../Shared/Layout.vue";
-import Build from "./pages/Build.vue";
-import Study from "./pages/Study.vue";
 import AppButton from "../../../components/AppButton.vue";
 import DeckFlashcard from "../../../components/DeckFlashcard.vue";
 import {route} from "ziggy-js";
 import {router} from "@inertiajs/vue3";
 import AppTip from "../../../components/AppTip.vue";
 import LoadingSpinner from "../../../Shared/LoadingSpinner.vue";
+import QuizzerWindow from "./UI/QuizzerWindow.vue";
+import ScoreStats from "../../../components/ScoreStats.vue";
+import {useDeckStudyStore} from "./Stores/DeckStudyStore.js";
+
+const DeckStudyStore = useDeckStudyStore();
 
 const props = defineProps({
     mode: {type: String, default: 'build'},
-    step: {type: String, default: 'select'},
-    deck: Object,
-    terms: Array,
 });
 
 const mode = ref(props.mode);
 const decks = ref([]);
-const terms = ref(props.terms);
-const selectedDeck = ref(props.deck);
+const selectedDeck = ref(null);
 
 const toggleMode = async () => {
     selectedDeck.value = null;
@@ -54,6 +53,13 @@ const toggleSelectDeck = (index) => {
     } else {
         selectedDeck.value = decks.value[index];
     }
+
+    if (DeckStudyStore.data.deck?.id === decks.value[index].id) {
+        DeckStudyStore.data.deck = null;
+
+    } else {
+        DeckStudyStore.data.deck = decks.value[index];
+    }
 }
 
 const toBuild = () => {
@@ -63,13 +69,21 @@ const toBuild = () => {
 };
 
 const toStudy = () => {
-    selectedDeck.value?.id
-    && router.get(route('deck-master.study', selectedDeck.value?.id));
+    if (DeckStudyStore.data.deck.id) {
+        router.get(route('deck-master.study', DeckStudyStore.data.deck.id));
+    }
 };
 
 onMounted(async () => {
     await fetchDecks(props.mode);
     selectedDeck.value = props.deck;
+
+    DeckStudyStore.reset();
+    DeckStudyStore.data.step = 'select';
+});
+
+onBeforeUnmount(() => {
+    DeckStudyStore.reset();
 });
 
 defineOptions({
@@ -79,7 +93,7 @@ defineOptions({
 
 <template>
     <Head title="Deck Master"/>
-    <div id="app-head" v-if="step === 'select'">
+    <div id="app-head">
         <h1>Deck Master</h1>
         <div id="app-nav">
             <div @click="toggleMode" id="app-mode-toggle" :class="mode">
@@ -89,37 +103,37 @@ defineOptions({
     </div>
 
     <div id="app-body" :class="{ 'dm-mode-study': mode === 'study' }">
-        <div id="dm-select" v-if="step === 'select'">
-            <div class="app-nav-interact">
+        <div id="dm-select">
+            <QuizzerWindow v-if="mode === 'study'">
+                <ScoreStats :model="DeckStudyStore.data.deck"/>
+                <div v-if="DeckStudyStore.data.deck" class="window-footer">
+                    <button @click="toStudy">
+                        Select Deck
+                    </button>
+                </div>
+            </QuizzerWindow>
+
+            <div class="app-nav-interact" v-if="mode === 'build'">
                 <div class="app-nav-interact-buttons">
-                    <AppButton v-if="mode === 'build'" @click="toBuild"
-                               :label="selectedDeck ? 'Edit' : 'New'"/>
-                    <AppButton v-if="mode === 'study'" @click="toStudy"
-                               label="Start" :disabled="!selectedDeck"/>
+                    <AppButton @click="toBuild" :label="selectedDeck ? 'Edit' : 'New'"/>
                 </div>
             </div>
 
-            <div v-if="!isLoading && decks.length > 0" class="deck-item-grid">
+            <div v-if="!isLoading && decks.length" class="deck-item-grid">
                 <DeckFlashcard v-for="(deck, index) in decks" :key="deck.id" :model="deck"
-                               :disabled="mode === 'study' && deck.terms.length === 0"
-                               :active="selectedDeck?.id === deck.id"
+                               :disabled="mode === 'study' && !deck.terms.length"
+                               :active="selectedDeck?.id === deck.id || selectedDeck?.id === deck.id"
                                @flip="toggleSelectDeck(index)"
                 />
             </div>
-            <div v-else-if="!isLoading && decks.length === 0" class="deck-item-grid">
+            <div v-else-if="!isLoading && !decks.length" class="deck-item-grid">
                 <AppTip>
-                    <p v-if="mode === 'build'">It looks like you haven't created any Decks yet. Click <b>New</b> to get started!</p>
+                    <p v-if="mode === 'build'">It looks like you haven't created any Decks yet. Click <b>New</b> to get
+                        started!</p>
                     <p v-if="mode === 'study'">It looks like you haven't pinned any Decks yet. Watch this space.</p>
                 </AppTip>
             </div>
             <LoadingSpinner v-show="isLoading"/>
-        </div>
-
-        <div id="dm-build" v-if="step === 'build'">
-            <Build :deck="selectedDeck"/>
-        </div>
-        <div id="dm-study" v-if="step === 'study'">
-            <Study :deck="selectedDeck" :terms="terms"/>
         </div>
     </div>
 </template>
