+++ resources/js/Pages/Workbench/DeckMaster/Pages/Practice.vue
@@ -1,0 +1,200 @@
+<script setup>
+import {onMounted, onUnmounted, reactive, ref, watch} from "vue";
+import {Carousel, Pagination, Slide} from "vue3-carousel";
+import 'vue3-carousel/dist/carousel.css';
+import TermFlashcard from "../UI/TermFlashcard.vue";
+import PopupWindow from "../../../../components/Modals/PopupWindow.vue";
+import WindowSection from "../../../../components/WindowSection.vue";
+import TermItem from "../../../../components/TermItem.vue";
+import ToggleSingle from "../../../../components/ToggleSingle.vue";
+import ToggleDouble from "../../../../components/ToggleDouble.vue";
+import QuizzerWindow from "../UI/QuizzerWindow.vue";
+import Layout from "../../../../Shared/Layout.vue";
+import {useDeckStudyStore} from "../Stores/DeckStudyStore.js";
+import LoadingSpinner from "../../../../Shared/LoadingSpinner.vue";
+
+const DeckStudyStore = useDeckStudyStore();
+
+defineOptions({
+    layout: Layout
+});
+
+const cards = reactive([]);
+const carouselRef = ref(null);
+const currentSlideIndex = ref(0);
+const isSliding = ref(false);
+const flipDefault = ref(false);
+const flipDefaultInflections = ref(false);
+const showTerm = ref(true);
+const showTranslit = ref(false);
+
+const handleSlideStart = () => {
+    isSliding.value = true;
+};
+
+const handleSlideEnd = ({currentSlideIndex: newIndex}) => {
+    currentSlideIndex.value = newIndex;
+    isSliding.value = false;
+};
+
+const toStart = () => {
+    carouselRef.value?.slideTo(0);
+
+    const flashcards = document.querySelectorAll('.term-flashcard');
+    flashcards.forEach(card => {
+        if (flipDefault.value) {
+            card.classList.add('flipped');
+        } else {
+            card.classList.remove('flipped');
+        }
+    });
+}
+
+const resetCards = () => {
+    cards.splice(0, cards.length, ...DeckStudyStore.data.terms);
+    currentSlideIndex.value = 0;
+    toStart();
+};
+
+const shuffleCards = () => {
+    const shuffled = [...cards].sort(() => Math.random() - 0.5);
+    cards.splice(0, cards.length, ...shuffled);
+    currentSlideIndex.value = 0;
+    toStart();
+};
+
+const handleGlobalKeydown = (event) => {
+    if (!carouselRef.value || isSliding.value) return;
+
+    switch (event.key) {
+        case 'ArrowRight':
+        case 'd':
+            carouselRef.value?.next();
+            break;
+        case 'ArrowLeft':
+        case 'a':
+            carouselRef.value?.prev();
+            break;
+        case ' ':
+            event.preventDefault();
+            flipDefault.value = !flipDefault.value;
+            break;
+    }
+};
+
+onMounted(() => {
+    cards.splice(0, cards.length, ...DeckStudyStore.data.terms);
+    window.addEventListener('keydown', handleGlobalKeydown);
+});
+
+onUnmounted(() => {
+    window.removeEventListener('keydown', handleGlobalKeydown);
+});
+
+watch(() => DeckStudyStore.data.terms, (newVal) => {
+    cards.splice(0, cards.length, ...newVal);
+    currentSlideIndex.value = 0;
+    toStart();
+})
+</script>
+
+<template>
+    <div id="dm-study">
+        <!--    <div class="window-container">-->
+        <!--        <div class="window-header">-->
+        <!--            <Link :href="route('deck-master.index', {mode: 'study'})" class="material-symbols-rounded">-->
+        <!--                arrow_back-->
+        <!--            </Link>-->
+        <!--            <div class="window-header-url">www.palweb.app/workbench/deck-master/study/{deck}</div>-->
+        <!--            <button class="material-symbols-rounded" @click="shuffleCards">shuffle</button>-->
+        <!--            <button class="material-symbols-rounded" @click="resetCards">undo</button>-->
+        <!--        </div>-->
+
+        <!--        <div class="window-section-head">-->
+        <!--            <h1>deck</h1>-->
+        <!--            <PinButton modelType="deck" :model="deck"/>-->
+        <!--            <DeckActions :model="deck"/>-->
+        <!--        </div>-->
+        <!--        <div class="window-content-head">-->
+        <!--            <div class="window-content-head-title">{{ deck.name }}</div>-->
+        <!--        </div>-->
+        <!--    </div>-->
+        <QuizzerWindow>
+            <div class="window-section-head">
+                <h2>Options</h2>
+                <PopupWindow title="Deck Master">
+                    <div>Options</div>
+                    <ul>
+                        <li><b>Reset</b> — Restores the Cards in the Deck to their original order.</li>
+                        <li><b>Shuffle</b> — Shuffles the Cards in the Deck.</li>
+                        <li><b>Default Side Toggle</b> — Toggles the default face of Cards in the Deck. Cards that
+                            have
+                            already been flipped will be unaffected.
+                        </li>
+                        <li><b>Show Inflections Toggle</b> — Toggles whether Inflections are shown on the Term side
+                            or
+                            on
+                            the Gloss side of the Card.
+                        </li>
+                        <li><b>Show Transcription</b> — Show the default transcription of the Term & its
+                            Inflections.
+                        </li>
+                        <li><b>Show Term</b> — Show the Term on the Gloss side of the Card.</li>
+                    </ul>
+                    <div>Keyboard Controls</div>
+                    <ul>
+                        <li><b>Left</b> / <b>A</b> — Previous Card</li>
+                        <li><b>Right</b> / <b>D</b> — Next Card</li>
+                        <li><b>Down</b> / <b>S</b> — Flip Card</li>
+                        <li><b>Space</b> — Toggle Default Side</li>
+                    </ul>
+                </PopupWindow>
+            </div>
+            <div class="quiz-settings-wrapper">
+                <ToggleDouble v-model="flipDefault" label="Initial Face" option-a="Term" option-b="Gloss"/>
+                <ToggleDouble v-model="flipDefaultInflections" label="Inflections" option-a="Back"
+                              option-b="Front"/>
+                <ToggleSingle v-model="showTranslit" label="Show Transcription"/>
+                <ToggleSingle v-model="showTerm" label="Show Term (Back)"/>
+            </div>
+            <WindowSection :visible="false">
+                <template #title>
+                    <h2>term</h2>
+                </template>
+                <template #content>
+                    <div class="model-list index-list">
+                        <TermItem :model="cards[currentSlideIndex]"
+                                  :glossId="cards[currentSlideIndex].deckPivot.gloss_id ?? null"/>
+                    </div>
+                </template>
+            </WindowSection>
+        </QuizzerWindow>
+
+        <Carousel
+            v-if="!DeckStudyStore.data.isLoading"
+            :items-to-show="1"
+            :wrap-around="true"
+            ref="carouselRef"
+            @slide-start="handleSlideStart"
+            @slide-end="handleSlideEnd"
+        >
+            <template #slides>
+                <Slide v-for="(term, index) in cards" :key="term.id">
+                    <TermFlashcard
+                        :model="term"
+                        :active="index === currentSlideIndex && !isSliding"
+                        :flipDefault="flipDefault"
+                        :showTerm="showTerm"
+                        :showTranslit="showTranslit"
+                        :flipDefaultInflections="flipDefaultInflections"
+                    />
+                </Slide>
+            </template>
+            <template #addons>
+                <Pagination/>
+                <div class="carousel-index">{{ currentSlideIndex + 1 }} out of {{ cards.length }}</div>
+            </template>
+        </Carousel>
+        <LoadingSpinner v-else/>
+    </div>
+</template>
