+++ resources/js/Pages/Workbench/DeckMaster/Pages/Quiz.vue
@@ -1,5 +1,5 @@
 <script setup>
-import {useQuizzerStore} from "../Stores/QuizzerStore.js";
+import {useDeckStudyStore} from "../Stores/DeckStudyStore.js";
 import {computed, ref} from "vue";
 import QuestionSelectGloss from "../UI/QuestionSelectGloss.vue";
 import AppTip from "../../../../components/AppTip.vue";
@@ -10,10 +10,10 @@ import QuizzerWindow from "../UI/QuizzerWindow.vue";
 import QuestionSelectTerm from "../UI/QuestionSelectTerm.vue";
 import LoadingSpinner from "../../../../Shared/LoadingSpinner.vue";
 
-const QuizzerStore = useQuizzerStore();
+const DeckStudyStore = useDeckStudyStore();
 
 const isValidRequest = computed(() => {
-    return !QuizzerStore.quiz.some(question => !question.response)
+    return !DeckStudyStore.quiz.some(question => !question.response)
 });
 
 const showInflections = ref(false);
@@ -25,43 +25,43 @@ const showTranslit = ref(false);
             <h2>Quiz</h2>
         </div>
         <AppTip>
-            <p v-if="QuizzerStore.settings.quizType === 'term-gloss'">
+            <p v-if="DeckStudyStore.settings.quizType === 'term-gloss'">
                 Select the meaning of the Arabic term in English.</p>
-            <p v-else-if="QuizzerStore.settings.quizType === 'term-inflection'">
+            <p v-else-if="DeckStudyStore.settings.quizType === 'term-inflection'">
                 Write the indicated inflection of the Term in Arabic.</p>
-            <p v-else-if="QuizzerStore.settings.quizType === 'sentence-term'">
+            <p v-else-if="DeckStudyStore.settings.quizType === 'sentence-term'">
                 Select the Term that best fits the blank in the Sentence. <b>Terms are listed in their Dictionary form,
                 not necessarily as they would be expected to appear in the Sentence.</b></p>
         </AppTip>
         <div class="quiz-settings-wrapper" style="justify-content: space-around">
             <ToggleSingle v-model="showTranslit" label="Show Transcription"/>
-            <ToggleSingle v-if="QuizzerStore.settings.quizType === 'term-gloss'"
+            <ToggleSingle v-if="DeckStudyStore.settings.quizType === 'term-gloss'"
                           v-model="showInflections" label="Show Inflections"/>
         </div>
     </QuizzerWindow>
 
-    <div class="quiz-container" v-if="!QuizzerStore.data.isLoading">
-        <QuestionSelectGloss v-if="QuizzerStore.settings.quizType === 'term-gloss'"
-                             v-for="(question, index) in QuizzerStore.quiz"
+    <div class="quiz-container" v-if="!DeckStudyStore.data.isLoading">
+        <QuestionSelectGloss v-if="DeckStudyStore.settings.quizType === 'term-gloss'"
+                             v-for="(question, index) in DeckStudyStore.quiz"
                              :question="question"
                              :index="index"
                              :showTranslit="showTranslit"
                              :showInflections="showInflections"
         />
-        <QuestionInputInflection v-if="QuizzerStore.settings.quizType === 'term-inflection'"
-                                 v-for="(question, index) in QuizzerStore.quiz"
+        <QuestionInputInflection v-if="DeckStudyStore.settings.quizType === 'term-inflection'"
+                                 v-for="(question, index) in DeckStudyStore.quiz"
                                  :question="question"
                                  :index="index"
                                  :showTranslit="showTranslit"
         />
-        <QuestionSelectTerm v-if="QuizzerStore.settings.quizType === 'sentence-term'"
-                            v-for="(question, index) in QuizzerStore.quiz"
+        <QuestionSelectTerm v-if="DeckStudyStore.settings.quizType === 'sentence-term'"
+                            v-for="(question, index) in DeckStudyStore.quiz"
                             :question="question" :index="index"
                             :showTranslit="showTranslit"
         />
     </div>
     <LoadingSpinner v-else/>
 
-    <AppButton :disabled="!isValidRequest" @click="QuizzerStore.submitQuiz" label="Submit"
+    <AppButton :disabled="!isValidRequest" @click="DeckStudyStore.submitQuiz" label="Submit"
                style="margin-block-end: 3.2rem"/>
 </template>
