+++ resources/js/Pages/Workbench/DeckMaster/UI/TermFlashcard.vue
